requirements.txt
pandas
numpy
scikit-learn
matplotlib
sentence-transformers
torch
umap-learn